refactor: refactored cards class-components to functions components
