+++ src/types/types.tsx
@@ -7,6 +7,20 @@ type User = {
   avatar: string;
 };
 
+type Card = {
+  id: number;
+  title: string;
+  description: string;
+  price: number;
+  discountPercentage: number;
+  rating: number;
+  stock: number;
+  brand: string;
+  category: string;
+  thumbnail: string;
+  images: string[];
+};
+
 type UserProps = {
   name: string;
   birth: string;
